Move noroute store to it's own package
