+++ noroute/no_route.go
@@ -1,4 +1,4 @@
-package route
+package noroute
 
 import (
     "log"
